revert: rename of _simple_cycle.py
