+++ autora/cycle/__init__.py
@@ -1,1 +1,1 @@
-from .simple import SimpleCycle as Cycle
+from ._simple_cycle import SimpleCycle as Cycle
